+++ yosai_intel_dashboard/src/adapters/ui/state/index.ts
@@ -1,5 +1,6 @@
 export * from './store';
 export * from './sessionSlice';
 export * from './analyticsSlice';
+export * from './uiSlice';
 
 export * from './ZustandProvider';